requirements.txt
Django==1.11.11
djangorestframework==3.7.7
Markdown==2.6.11
pytz==2018.3
psycopg2
